firmware: add basic goto ra functionality
This adds the basic implementation to execute automatic movement
to the target location based on the current input location.
As a cli command the user can enter the current focues star
location and the target star location which instructs the tracker
to move into the correct direction.
cli syntax is as follows:
gotoRA 12° 13' 14.3" 21° 14' 15.4" except the extra characters are
not supported in the cli so use the following:
gotoRA 12 13 14.3 21 14 15.4